packages/markdowndeck/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markdowndeck"
version = "0.1.2"
description = "Convert Markdown to Google Slides presentations"
authors = [
    {name = "Arclio Team", email = "info@arclio.com"},
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["markdown", "slides", "presentations", "google slides"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "markdown-it-py>=3.0.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "requests>=2.28.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["src/markdowndeck"]

[tool.hatch.metadata]
allow-direct-references = true
